Restructure package into subpackages
