+++ docker_ctp/__main__.py
@@ -1,0 +1,6 @@
+"""Run docker-ctp as a module."""
+
+from .main import main
+
+if __name__ == "__main__":  # pragma: no cover
+    main()
