+++ docker_ctp/utils/auth.py
