+++ docker_ctp/utils/config_generation.py
