+++ docker_ctp/utils/dependency_checker.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import logging
 import shutil
 import subprocess
