+++ docker_ctp/utils/input_validation.py
